fix: move QuotaRepo to db.py
